Move selection Pinia store under stores/ui/ to mirror old Vuex layout
Move src/stores/selection.ts to src/stores/ui/selection.ts and update all import paths to reflect the new location. The store ID remains unchanged.
