+++ src/App.vue
@@ -86,7 +86,7 @@ import DiceRoller from "~/components/ui/generic/DiceRoller"
 import GameMenu from "~/components/ui/menus/GameMenu"
 import SystemMenu from "~/components/ui/menus/SystemMenu"
 import { usePreferencesStore } from "~/stores/ui/preferences"
-import { useSelectionStore, View } from "~/stores/selection"
+import { useSelectionStore, View } from "~/stores/ui/selection"
 
 defineOptions({ name: "App" })
