+++ src/components/game/battle/StrikeConfirmation.vue
@@ -66,7 +66,7 @@ import { computed } from "vue"
 import { isEqual, range } from "lodash-es"
 import { useStore } from "vuex"
 import { BattleCreature, isRangestrike, Strike } from "~/models/battle"
-import { useSelectionStore } from "~/stores/selection"
+import { useSelectionStore } from "~/stores/ui/selection"
 
 defineOptions({ name: "StrikeConfirmation" })
