+++ src/components/game/masterboard/MasterboardStack.vue
@@ -69,7 +69,7 @@ import { MasterboardPhase, Path } from "~/models/game"
 import masterboard, { HexEdge, MasterboardEdge, MasterboardHex } from "~/models/masterboard"
 import { Player } from "~/models/player"
 import { Stack } from "~/models/stack"
-import { useSelectionStore } from "~/stores/selection"
+import { useSelectionStore } from "~/stores/ui/selection"
 import { Transformation, Transformations, TransformationType } from "~/utils/svg"
 import EngageIcon from "../../ui/game/EngageIcon.vue"
 import Creature from "../Creature.vue"
