+++ src/components/ui/game/StackPanel.vue
@@ -114,7 +114,7 @@ import { MasterboardPhase } from "~/models/game"
 import masterboard, { MasterboardHex, Terrain } from "~/models/masterboard"
 import { Player } from "~/models/player"
 import { MusterChoice, MusterPossibility, Stack } from "~/models/stack"
-import { useSelectionStore } from "~/stores/selection"
+import { useSelectionStore } from "~/stores/ui/selection"
 import { mod } from "~/utils/math"
 import Creature from "../../game/Creature.vue"
 import PlayerMarker from "../../game/Marker.vue"
